new：添加AMemory缓存机制 update :调整BaseTBActivity new：添加UserVM new：添加上传图片界面、保存图片功能

+++ app/src/main/java/com/iblueroad/hyt/data/bmob/model/BMUser.kt
@@ -7,7 +7,7 @@ import cn.bmob.v3.BmobUser
  */
 
 data class BMUser(
-    var uid: Int = 0,
+    var uid: Long = 0L,
     var sex: Int = 0,//0:男 1：女
     var labels: List<String>? = null,
     var carrer: String? = null,
